% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs4_book.R
\name{bs4_book}
\alias{bs4_book}
\title{bs4_book}
\usage{
bs4_book(
  copy_supporting_files = TRUE,
  theme = bookdown::bs4_book_theme(primary = "#086a6a", base_font =
    bslib::font_google("Lexend"), heading_font = bslib::font_google("Roboto"), code_font
    = bslib::font_google("JetBrains Mono"), ),
  template = system.file("rmarkdown/resources/html_files/bs4_book_ptbr.html", package =
    "fnaufelRmd"),
  includes = rmarkdown::includes(in_header = "html_files/preamble.html"),
  css = "html_files/styles.css",
  repo = NULL,
  site = bookdown::bookdown_site,
  lib_dir = "libs",
  pandoc_args = NULL,
  extra_dependencies = NULL,
  split_bib = FALSE,
  section_divs = TRUE,
  fig_width = 7,
  fig_height = 5,
  fig_retina = 2,
  fig_caption = TRUE,
  dev = "png",
  df_print = "paged",
  mathjax = "default",
  keep_md = FALSE,
  md_extensions = NULL,
  ...
)
}
\arguments{
\item{copy_supporting_files}{If TRUE, copy supporting files (css, js, html) to document dir before rendering}

\item{theme}{see \link[bookdown:bs4_book]{bookdown::bs4_book}}

\item{template}{see \link[bookdown:bs4_book]{bookdown::bs4_book}}

\item{includes}{see \link[bookdown:bs4_book]{bookdown::bs4_book}}

\item{css}{see \link[bookdown:bs4_book]{bookdown::bs4_book}}

\item{repo}{see \link[bookdown:bs4_book]{bookdown::bs4_book}}

\item{site}{see \link[bookdown:bs4_book]{bookdown::bs4_book}}

\item{lib_dir}{see \link[bookdown:bs4_book]{bookdown::bs4_book}}

\item{pandoc_args}{see \link[bookdown:bs4_book]{bookdown::bs4_book}}

\item{extra_dependencies}{see \link[bookdown:bs4_book]{bookdown::bs4_book}}

\item{split_bib}{see \link[bookdown:bs4_book]{bookdown::bs4_book}}

\item{section_divs}{see \link[bookdown:bs4_book]{bookdown::bs4_book}}

\item{fig_width}{see \link[bookdown:bs4_book]{bookdown::bs4_book}}

\item{fig_height}{see \link[bookdown:bs4_book]{bookdown::bs4_book}}

\item{fig_retina}{see \link[bookdown:bs4_book]{bookdown::bs4_book}}

\item{fig_caption}{see \link[bookdown:bs4_book]{bookdown::bs4_book}}

\item{dev}{see \link[bookdown:bs4_book]{bookdown::bs4_book}}

\item{df_print}{see \link[bookdown:bs4_book]{bookdown::bs4_book}}

\item{mathjax}{see \link[bookdown:bs4_book]{bookdown::bs4_book}}

\item{keep_md}{see \link[bookdown:bs4_book]{bookdown::bs4_book}}

\item{md_extensions}{see \link[bookdown:bs4_book]{bookdown::bs4_book}}

\item{...}{see \link[bookdown:bs4_book]{bookdown::bs4_book}}
}
\value{
see \link[bookdown:bs4_book]{bookdown::bs4_book}
}
\description{
Custom HTML template
}
\details{
Loads additional style and configs
}
\author{
fnaufel
}
