% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_if_stale.R
\name{copy_if_stale}
\alias{copy_if_stale}
\title{copy_if_stale}
\usage{
copy_if_stale(from, to)
}
\arguments{
\item{from}{absolute path of single file or single dir to copy from}

\item{to}{absolute path of single \strong{dir} to copy to}
}
\description{
Copy files from \code{from} to \code{to} if the former are
more recent than the latter. All directories in \code{from} are copied
unconditionally.
}
