% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_if_stale.R
\name{copy_if_stale_single}
\alias{copy_if_stale_single}
\title{copy_if_stale_single}
\usage{
copy_if_stale_single(from, to)
}
\arguments{
\item{from}{absolute path of single \strong{file} or \strong{dir} to copy from}

\item{to}{absolute path of single \strong{file} or \strong{dir} to copy to}
}
\description{
copy_if_stale_single
}
