% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_report.R
\name{html_report}
\alias{html_report}
\title{html_report}
\usage{
html_report(
  toc = TRUE,
  toc_depth = 3,
  toc_float = FALSE,
  number_sections = TRUE,
  anchor_sections = TRUE,
  section_divs = TRUE,
  fig_width = 7,
  fig_height = 5,
  fig_retina = 2,
  fig_caption = TRUE,
  dev = "png",
  df_print = "paged",
  code_folding = "show",
  code_download = TRUE,
  self_contained = TRUE,
  theme = bslib::bs_theme(primary = "#086a6a", base_font =
    bslib::font_google("Lexend"), heading_font = bslib::font_google("Roboto"), code_font
    = bslib::font_google("JetBrains Mono"), ),
  highlight = "default",
  mathjax = "default",
  template = "default",
  extra_dependencies = NULL,
  css = "html/styles.css",
  includes = rmarkdown::includes(in_header = "html/preamble.html"),
  keep_md = FALSE,
  lib_dir = "libs",
  md_extensions = NULL,
  pandoc_args = NULL,
  ...
)
}
\arguments{
\item{toc}{see \link[rmarkdown:html_document]{rmarkdown::html_document}}

\item{dev}{see \link[rmarkdown:html_document]{rmarkdown::html_document}}

\item{theme}{see \link[rmarkdown:html_document]{rmarkdown::html_document}}

\item{highlight}{see \link[rmarkdown:html_document]{rmarkdown::html_document}}

\item{mathjax}{see \link[rmarkdown:html_document]{rmarkdown::html_document}}

\item{template}{see \link[rmarkdown:html_document]{rmarkdown::html_document}}

\item{css}{see \link[rmarkdown:html_document]{rmarkdown::html_document}}

\item{includes}{see \link[rmarkdown:html_document]{rmarkdown::html_document}}

\item{...}{see \link[rmarkdown:html_document]{rmarkdown::html_document}}

\item{number}{see \link[rmarkdown:html_document]{rmarkdown::html_document}}

\item{anchor}{see \link[rmarkdown:html_document]{rmarkdown::html_document}}

\item{section}{see \link[rmarkdown:html_document]{rmarkdown::html_document}}

\item{fig}{see \link[rmarkdown:html_document]{rmarkdown::html_document}}

\item{df}{see \link[rmarkdown:html_document]{rmarkdown::html_document}}

\item{code}{see \link[rmarkdown:html_document]{rmarkdown::html_document}}

\item{self}{see \link[rmarkdown:html_document]{rmarkdown::html_document}}

\item{extra}{see \link[rmarkdown:html_document]{rmarkdown::html_document}}

\item{keep}{see \link[rmarkdown:html_document]{rmarkdown::html_document}}

\item{lib}{see \link[rmarkdown:html_document]{rmarkdown::html_document}}

\item{md}{see \link[rmarkdown:html_document]{rmarkdown::html_document}}

\item{pandoc}{see \link[rmarkdown:html_document]{rmarkdown::html_document}}
}
\value{
see \link[rmarkdown:html_document]{rmarkdown::html_document}
}
\description{
Custom HTML template
}
\details{
Loads additional style and configs
}
\author{
fnaufel
}
