% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_named_lists.R
\name{merge_named_lists}
\alias{merge_named_lists}
\title{merge_named_lists}
\usage{
merge_named_lists(l1, l2)
}
\arguments{
\item{l1}{one list with named elements}

\item{l2}{another list with named elements}
}
\value{
merged list
}
\description{
merge_named_lists
}
\author{
Fernando Naufel
}
