% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_book.R
\name{new_book}
\alias{new_book}
\title{new_book}
\usage{
new_book(name = "index.Rmd")
}
\arguments{
\item{name}{Main file name, Default: 'index.Rmd'}
}
\value{
The file name of the new document (invisibly)
}
\description{
Creates new files based on the fnaufelRmd::book template
}
\author{
fnaufel
}
