% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_report.R
\name{new_report}
\alias{new_report}
\title{new_report}
\usage{
new_report(name = "index.Rmd")
}
\arguments{
\item{name}{Main file name, Default: 'index.Rmd'}
}
\value{
The file name of the new document (invisibly)
}
\description{
Creates new file based on the fnaufelRmd::report template
}
\author{
fnaufel
}
