% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_book.R
\name{pdf_book}
\alias{pdf_book}
\title{pdf_book}
\usage{
pdf_book(
  copy_supporting_files = TRUE,
  toc = TRUE,
  number_sections = TRUE,
  fig_caption = TRUE,
  pandoc_args = c("--variable=link-as-notes", paste0("--variable=", "resources:",
    system.file("rmarkdown/resources", package = "fnaufelRmd")),
    "--variable=subparagraph", paste0("--filter=",
    system.file("rmarkdown/resources/pandoc_filters/dollar.py3", package =
    "fnaufelRmd")), paste0("--filter=",
    system.file("rmarkdown/resources/pandoc_filters/hl.py3", package = "fnaufelRmd"))),
  base_format = rmarkdown::pdf_document,
  toc_unnumbered = TRUE,
  toc_appendix = TRUE,
  toc_bib = TRUE,
  quote_footer = NULL,
  highlight_bw = FALSE,
  df_print = "tibble",
  keep_tex = TRUE,
  latex_engine = "xelatex",
  citation_package = "default",
  includes = rmarkdown::includes(in_header =
    system.file("rmarkdown/resources/tex/bd-preamble.tex", package = "fnaufelRmd")),
  md_extensions = "+fenced_divs+bracketed_spans",
  template = system.file("rmarkdown/resources/tex/bd-format.latex", package =
    "fnaufelRmd"),
  ...
)
}
\arguments{
\item{copy_supporting_files}{If TRUE, copy supporting files (.csl etc.) to document dir before rendering}

\item{toc}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{number_sections}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{fig_caption}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{pandoc_args}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{base_format}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{toc_unnumbered}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{toc_appendix}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{toc_bib}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{quote_footer}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{highlight_bw}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{df_print}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{keep_tex}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{latex_engine}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{citation_package}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{includes}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{md_extensions}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{template}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{...}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}
}
\value{
see \link[bookdown:pdf_book]{bookdown::pdf_book}
}
\description{
Custom PDF template for books
}
\details{
Loads additional style and configs
}
\author{
fnaufel
}
