% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_report.R
\name{pdf_report}
\alias{pdf_report}
\title{pdf_report}
\usage{
pdf_report(
  copy_supporting_files = TRUE,
  toc = TRUE,
  toc_depth = 2,
  number_sections = TRUE,
  fig_width = 6.5,
  fig_height = 4.5,
  fig_crop = "auto",
  fig_caption = TRUE,
  dev = "pdf",
  df_print = "tibble",
  highlight = "default",
  keep_tex = TRUE,
  keep_md = FALSE,
  latex_engine = "xelatex",
  citation_package = "natbib",
  includes = rmarkdown::includes(in_header =
    system.file("rmarkdown/resources/tex/report-preamble.tex", package = "fnaufelRmd")),
  md_extensions = "+fenced_divs+bracketed_spans",
  template = system.file("rmarkdown/resources/tex/report-format.latex", package =
    "fnaufelRmd"),
  output_extensions = NULL,
  pandoc_args = c("--variable=link-as-notes", paste0("--variable=", "resources:",
    system.file("rmarkdown/resources", package = "fnaufelRmd")),
    "--variable=subparagraph", paste0("--filter=",
    system.file("rmarkdown/resources/pandoc_filters/dollar.py3", package =
    "fnaufelRmd")), paste0("--filter=",
    system.file("rmarkdown/resources/pandoc_filters/hl.py3", package = "fnaufelRmd"))),
  extra_dependencies = NULL,
  ...
)
}
\arguments{
\item{copy_supporting_files}{If TRUE, copy supporting files (.csl etc.) to document dir before rendering}

\item{toc}{see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}

\item{toc_depth}{see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}

\item{number_sections}{see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}

\item{fig_width}{see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}

\item{fig_height}{see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}

\item{fig_crop}{see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}

\item{fig_caption}{see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}

\item{dev}{see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}

\item{df_print}{see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}

\item{highlight}{see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}

\item{keep_tex}{see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}

\item{keep_md}{see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}

\item{latex_engine}{see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}

\item{citation_package}{see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}

\item{includes}{see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}

\item{md_extensions}{see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}

\item{template}{see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}

\item{output_extensions}{see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}

\item{pandoc_args}{see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}

\item{extra_dependencies}{see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}

\item{...}{see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}
}
\value{
see \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}
}
\description{
Custom PDF template for reports
}
\details{
Loads additional style and configs
}
\author{
fnaufel
}
