% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_report.R
\name{html_report}
\alias{html_report}
\title{html_report}
\usage{
html_report(
  copy_supporting_files = TRUE,
  toc = TRUE,
  toc_depth = 3,
  toc_float = FALSE,
  number_sections = TRUE,
  anchor_sections = TRUE,
  section_divs = TRUE,
  fig_width = 7,
  fig_height = 5,
  fig_retina = 2,
  fig_caption = TRUE,
  dev = "png",
  df_print = "paged",
  code_folding = "none",
  code_download = TRUE,
  self_contained = TRUE,
  theme = "default",
  highlight = "default",
  mathjax = "default",
  template = "default",
  extra_dependencies = NULL,
  css = "default",
  includes = NULL,
  keep_md = FALSE,
  lib_dir = "libs",
  md_extensions = NULL,
  pandoc_args = NULL,
  highlight_downlit = TRUE,
  ...
)
}
\arguments{
\item{copy_supporting_files}{If TRUE, copy supporting files (css, js, html) to document dir before rendering}

\item{toc}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{toc_depth}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{toc_float}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{number_sections}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{anchor_sections}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{section_divs}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{fig_width}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{fig_height}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{fig_retina}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{fig_caption}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{dev}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{df_print}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{code_folding}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{code_download}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{self_contained}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{theme}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{highlight}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{mathjax}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{template}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{extra_dependencies}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{css}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{includes}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{keep_md}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{lib_dir}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{md_extensions}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{pandoc_args}{see \link[bookdown:html_document2]{bookdown::html_document2}}

\item{...}{see \link[bookdown:html_document2]{bookdown::html_document2}}
}
\value{
see \link[bookdown:html_document2]{bookdown::html_document2}
}
\description{
Custom HTML template for reports
}
\details{
Loads additional style and configs
}
\author{
fnaufel
}
